Added mouseDragged
Looking for mouseDragged makes it a bit easier to see what's going on and finetune where one wants the origin point. Maybe the next step is to draw an origin point that one can move around as if its a point. This would allow the window to drop the "Capture Clicks" option, and prevent people from accidentally moving their origin if they click somewhere random. Or maybe it's just a modifier key...